# openapi/component-services.v1.yaml
openapi: 3.1.0
info:
  title: PACP Component Services API
  version: v1
  description: Internal distributed HTTP contracts between PACP components, runners, nodes, and operators.
servers:
  - url: "{catalog_url}"
    description: Configured catalog service base URL.
    variables:
      catalog_url:
        default: https://catalog.pacp.example
        description: Replace with the deployed catalog URL, or a local development URL.
  - url: "{jobs_url}"
    description: Configured jobs service base URL.
    variables:
      jobs_url:
        default: https://jobs.pacp.example
        description: Replace with the deployed jobs URL, or a local development URL.
  - url: "{leases_url}"
    description: Configured leases service base URL.
    variables:
      leases_url:
        default: https://leases.pacp.example
        description: Replace with the deployed leases URL, or a local development URL.
  - url: "{artifacts_url}"
    description: Configured artifacts service base URL.
    variables:
      artifacts_url:
        default: https://artifacts.pacp.example
        description: Replace with the deployed artifacts URL, or a local development URL.
  - url: "{policy_url}"
    description: Configured policy service base URL.
    variables:
      policy_url:
        default: https://policy.pacp.example
        description: Replace with the deployed policy URL, or a local development URL.
  - url: "{node_url}"
    description: Configured node service base URL.
    variables:
...
security:
  - componentBearer: []
...
paths:
...
  /v1/provider/manifest:
    get:
      tags: [provider]
      operationId: getProviderManifest
      x-operation-audience: [component, worker, agent]
      x-policy-action: provider.manifest.read
      security: []
      responses:
        "200":
          $ref: "#/components/responses/ProviderManifestResponse"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    componentBearer:
      type: http
      scheme: bearer
      description: Optional transport guard used by component services when configured.
    nodeBearer:
      type: http
      scheme: bearer
      description: Node-local bearer credential configured in the node config file.
    providerBearer:
      type: http
      scheme: bearer
      description: Optional bearer credential configured by provider deployments for invoke and content routes.
    runnerMonitorBearer:
      type: http
      scheme: bearer
      description: Optional bearer credential configured with pacp-runner -monitor-token.
...
  responses:
    Error:
      description: Error envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ProviderManifestResponse:
      description: Provider manifest envelope.
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/SuccessEnvelope"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ProviderManifest"
...
  schemas:
    AnyMap:
      type: object
      additionalProperties: true
    Links:
      type: object
      additionalProperties: true
    Meta:
      type: object
      required: [request_id, schema_version]
      properties:
        request_id:
          type: string
        schema_version:
          type: string
    SuccessEnvelope:
      type: object
      required: [ok, data, links, meta]
      properties:
        ok:
          const: true
        data: {}
        links:
          $ref: "#/components/schemas/Links"
        meta:
          $ref: "#/components/schemas/Meta"
    ErrorEnvelope:
      type: object
      required: [ok, error, links, meta]
      properties:
        ok:
          const: false
        error:
          $ref: "#/components/schemas/ErrorObject"
        links:
          $ref: "#/components/schemas/Links"
        meta:
          $ref: "#/components/schemas/Meta"
    ErrorObject:
      type: object
      required: [code, message, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        retryable:
          type: boolean
...
    ProviderManifest:
      type: object
      required: [schema_version, service, provider, capabilities]
      properties:
        schema_version:
          type: string
        service:
          $ref: "#/components/schemas/Service"
        provider:
          $ref: "#/components/schemas/Provider"
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/Capability"
...
    Service:
      type: object
      required: [id, name, description, version, provider_kind]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        version:
          type: string
        provider_kind:
          type: string
        tags:
          type: array
          items:
            type: string
    Provider:
      type: object
      required: [endpoint]
      properties:
        endpoint:
          type: string
          format: uri
          pattern: '^https?://[^?#]+$'
        node_id:
          type: string
        health_path:
          type: string
          pattern: '^/[^?#]*$'
    Capability:
      type: object
      required: [id, name, description, execution_mode, input_schema, output_schema, examples, side_effects, resource_hints, artifact_hints, timeout_hint]
      properties:
        id:
          type: string
        service_id:
          type: string
        name:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        execution_mode:
          type: string
          enum: [sync, async, either]
        input_schema:
          $ref: "#/components/schemas/AnyMap"
        output_schema:
          $ref: "#/components/schemas/AnyMap"
        examples:
          type: array
          items:
            $ref: "#/components/schemas/AnyMap"
        side_effects:
          type: string
          enum: [none, read, write, external, destructive]
        resource_hints:
          type: array
          items:
            $ref: "#/components/schemas/ResourceHint"
        artifact_hints:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactHint"
        timeout_hint:
          type: string
    ResourceHint:
      type: object
      required: [selector, required]
      properties:
        selector:
          type: string
        required:
          type: boolean
        quantity:
          type: integer
        notes:
          type: string
    ArtifactHint:
      type: object
      required: [media_type]
      properties:
        media_type:
          type: string
        count:
          type: string
        notes:
          type: string
...